refactor: modify ProductController as modified Product entity

+++ src/main/java/gift/controller/ProductController.java
@@ -12,6 +12,7 @@ import org.springframework.http.ResponseEntity;
 import org.springframework.validation.BindingResult;
 import org.springframework.web.bind.annotation.*;
 
+import java.util.Optional;
 import java.util.stream.Collectors;
 
 @RestController
@@ -39,32 +40,55 @@ public class ProductController {
     @PostMapping
     public ResponseEntity<?> addProduct(@Valid @RequestBody ProductRequest productRequest, BindingResult bindingResult) {
         if(bindingResult.hasErrors()) {
-            String errors = bindingResult.getAllErrors().stream()
-                    .map(error -> error.getDefaultMessage())
-                    .collect(Collectors.joining(", "));
+            String errors = bindingResult.getAllErrors().stream().map(error -> error.getDefaultMessage()).collect(Collectors.joining(","));
             return ResponseEntity.badRequest().body(errors);
         }
-        Product product = new Product();
-        product.setName(productRequest.getName());
-        product.setPrice(productRequest.getPrice());
-        product.setImageUrl(productRequest.getImageUrl());
-        return ResponseEntity.ok(productService.save(product));
+        Product product = new Product(
+                null,
+                productRequest.getName(),
+                productRequest.getPrice(),
+                productRequest.getImageUrl(),
+                null // Category 설정
+        );
+        return ResponseEntity.ok(new ProductResponse(
+                productService.save(product).getId(),
+                product.getName(),
+                product.getPrice(),
+                product.getImageUrl()
+        ));
     }
 
     @PutMapping("/{id}")
-    public ResponseEntity<?> updateProduct(@PathVariable Long id, @Valid @RequestBody ProductRequest productRequest, BindingResult bindingResult) {
+    public ResponseEntity<?> updateProduct(@PathVariable Long id, @RequestBody ProductRequest productRequest, BindingResult bindingResult) {
         if(bindingResult.hasErrors()) {
-            String errors = bindingResult.getAllErrors().stream()
-                    .map(error -> error.getDefaultMessage())
-                    .collect(Collectors.joining(", "));
+            String errors = bindingResult.getAllErrors().stream().map(error -> error.getDefaultMessage()).collect(Collectors.joining(","));
             return ResponseEntity.badRequest().body(errors);
         }
-        Product product = new Product();
-        return ResponseEntity.ok(productService.save(product));
+        Optional<Product> existingOpt = productService.findById(id);
+        if(existingOpt.isEmpty()) {
+            return ResponseEntity.notFound().build();
+        }
+        Product existingProduct = existingOpt.get();
+        existingProduct.update(
+          productRequest.getPrice(),
+          productRequest.getName(),
+          productRequest.getImageUrl(),
+          null // Category
+        );
+        return ResponseEntity.ok(new ProductResponse(
+                productService.save(existingProduct).getId(),
+                existingProduct.getName(),
+                existingProduct.getPrice(),
+                existingProduct.getImageUrl()
+        ));
     }
+
 
     @DeleteMapping("/{id}")
     public ResponseEntity<?> deleteProduct(@PathVariable Long id) {
+        if(productService.findById(id).isEmpty()) {
+            return ResponseEntity.notFound().build();
+        }
         productService.deleteById(id);
         return ResponseEntity.ok().build();
     }
